Agregando funcionalidades varias

+++ src/main/java/com/bootcamp/btmsaccounts/model/Credit.java
@@ -24,8 +24,8 @@ public class Credit {
     private String accountNumber; //Campo futuro, para validar que vaya asociada a una cuenta de ahorro
     private String status;
     private String customerId;
-    private double creditLimit; // Si se trata de tarjeta de crédito (Monto Original de la TC)
-    private double availableCredit; // Si se trata de tarjeta de crédito (Monto actual de la linea de credito)
+    private BigDecimal creditLimit; // Si se trata de tarjeta de crédito (Monto Original de la TC)
+    private BigDecimal availableCredit; // Si se trata de tarjeta de crédito (Monto actual de la linea de credito)
     private String creditSubType; // Credito/Prestamo(Personal, Empresarial), Tarjeta de credito
     private String idCreditCard; // Solo para TC
     private String lastTransactionId; // Solo TC - Campo para validar la ultima transaccion de la TC
